feat: install typescript
