+++ package.json
@@ -26,5 +26,10 @@
         "path": "./snippets/typescriptreactsnippets.code-snippets"
       }
     ]
+  },
+  "devDependencies": {
+    "@types/node": "^20.3.1",
+    "ts-node": "^10.9.1",
+    "typescript": "^5.1.3"
   }
 }
